SY-4653: Fold normalization specs into existing v0 spec files
