+++ core/pkg/service/ranger/alias/versions/v0/alias_test.go
@@ -15,6 +15,13 @@ import (
 	. "github.com/onsi/gomega"
 	channel "github.com/synnaxlabs/synnax/pkg/service/channel/versions/v0"
 	v0 "github.com/synnaxlabs/synnax/pkg/service/ranger/alias/versions/v0"
+	"github.com/synnaxlabs/x/encoding/msgpack"
+	"github.com/synnaxlabs/x/gorp"
+	gorptestutil "github.com/synnaxlabs/x/gorp/testutil"
+	"github.com/synnaxlabs/x/kv/memkv"
+	"github.com/synnaxlabs/x/migrate"
+	"github.com/synnaxlabs/x/query"
+	. "github.com/synnaxlabs/x/testutil"
 )
 
 var _ = Describe("Alias", func() {
@@ -31,4 +38,32 @@ var _ = Describe("Alias", func() {
 			Expect(v0.Alias{}.SetOptions()).To(BeNil())
 		})
 	})
+})
+
+var _ = Describe("NormalizeKeys", func() {
+	It(
+		"Should lift a Alias row stored under the pre-v0.54 key format",
+		func(ctx SpecContext) {
+			kvDB := memkv.New()
+			db := DeferClose(gorp.Wrap(kvDB, gorp.WithCodec(msgpack.Codec)))
+			e := v0.Alias{Range: uuid.New(), Channel: 65538, Alias: "pt"}
+			legacy := gorptestutil.SetPreV54Row(
+				ctx,
+				kvDB,
+				"Alias",
+				e.GorpKey(),
+				e,
+			)
+			table := MustOpen(gorp.OpenTable(ctx, gorp.TableConfig[string, v0.Alias]{
+				DB:         db,
+				Migrations: []migrate.Migration{v0.NormalizeKeys},
+			}))
+			var res v0.Alias
+			Expect(table.NewRetrieve().
+				Where(gorp.MatchKeys[string, v0.Alias](e.GorpKey())).
+				Entry(&res).Exec(ctx, db)).To(Succeed())
+			Expect(res).To(Equal(e))
+			Expect(db.Get(ctx, legacy)).Error().To(MatchError(query.ErrNotFound))
+		},
+	)
 })
