+++ core/pkg/service/ranger/kv/versions/v0/kv_test.go
@@ -14,6 +14,13 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v0 "github.com/synnaxlabs/synnax/pkg/service/ranger/kv/versions/v0"
+	"github.com/synnaxlabs/x/encoding/msgpack"
+	"github.com/synnaxlabs/x/gorp"
+	gorptestutil "github.com/synnaxlabs/x/gorp/testutil"
+	"github.com/synnaxlabs/x/kv/memkv"
+	"github.com/synnaxlabs/x/migrate"
+	"github.com/synnaxlabs/x/query"
+	. "github.com/synnaxlabs/x/testutil"
 )
 
 var _ = Describe("Pair", func() {
@@ -36,4 +43,32 @@ var _ = Describe("Pair", func() {
 			Expect(v0.Pair{}.CustomTypeName()).To(Equal("KVPair"))
 		})
 	})
+})
+
+var _ = Describe("NormalizeKeys", func() {
+	It(
+		"Should lift a Pair row stored under the pre-v0.54 key format",
+		func(ctx SpecContext) {
+			kvDB := memkv.New()
+			db := DeferClose(gorp.Wrap(kvDB, gorp.WithCodec(msgpack.Codec)))
+			e := v0.Pair{Range: uuid.New(), Key: "operator", Value: "ada"}
+			legacy := gorptestutil.SetPreV54Row(
+				ctx,
+				kvDB,
+				"Pair",
+				e.GorpKey(),
+				e,
+			)
+			table := MustOpen(gorp.OpenTable(ctx, gorp.TableConfig[string, v0.Pair]{
+				DB:         db,
+				Migrations: []migrate.Migration{v0.NormalizeKeys},
+			}))
+			var res v0.Pair
+			Expect(table.NewRetrieve().
+				Where(gorp.MatchKeys[string, v0.Pair](e.GorpKey())).
+				Entry(&res).Exec(ctx, db)).To(Succeed())
+			Expect(res).To(Equal(e))
+			Expect(db.Get(ctx, legacy)).Error().To(MatchError(query.ErrNotFound))
+		},
+	)
 })
