+++ core/pkg/service/schematic/versions/v0/schematic_test.go
@@ -14,6 +14,13 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v0 "github.com/synnaxlabs/synnax/pkg/service/schematic/versions/v0"
+	"github.com/synnaxlabs/x/encoding/msgpack"
+	"github.com/synnaxlabs/x/gorp"
+	gorptestutil "github.com/synnaxlabs/x/gorp/testutil"
+	"github.com/synnaxlabs/x/kv/memkv"
+	"github.com/synnaxlabs/x/migrate"
+	"github.com/synnaxlabs/x/query"
+	. "github.com/synnaxlabs/x/testutil"
 )
 
 var _ = Describe("Schematic", func() {
@@ -29,4 +36,34 @@ var _ = Describe("Schematic", func() {
 			Expect(v0.Schematic{}.SetOptions()).To(BeNil())
 		})
 	})
+})
+
+var _ = Describe("NormalizeKeys", func() {
+	It(
+		"Should lift a Schematic row stored under the pre-v0.54 key format",
+		func(ctx SpecContext) {
+			kvDB := memkv.New()
+			db := DeferClose(gorp.Wrap(kvDB, gorp.WithCodec(msgpack.Codec)))
+			e := v0.Schematic{Key: uuid.New(), Name: "Feed"}
+			legacy := gorptestutil.SetPreV54Row(
+				ctx,
+				kvDB,
+				"Schematic",
+				e.GorpKey(),
+				e,
+			)
+			table := MustOpen(
+				gorp.OpenTable(ctx, gorp.TableConfig[v0.Key, v0.Schematic]{
+					DB:         db,
+					Migrations: []migrate.Migration{v0.NormalizeKeys},
+				}),
+			)
+			var res v0.Schematic
+			Expect(table.NewRetrieve().
+				Where(gorp.MatchKeys[v0.Key, v0.Schematic](e.GorpKey())).
+				Entry(&res).Exec(ctx, db)).To(Succeed())
+			Expect(res).To(Equal(e))
+			Expect(db.Get(ctx, legacy)).Error().To(MatchError(query.ErrNotFound))
+		},
+	)
 })
