+++ core/pkg/service/status/versions/v0/status_test.go
@@ -13,6 +13,13 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v0 "github.com/synnaxlabs/synnax/pkg/service/status/versions/v0"
+	"github.com/synnaxlabs/x/encoding/msgpack"
+	"github.com/synnaxlabs/x/gorp"
+	gorptestutil "github.com/synnaxlabs/x/gorp/testutil"
+	"github.com/synnaxlabs/x/kv/memkv"
+	"github.com/synnaxlabs/x/migrate"
+	"github.com/synnaxlabs/x/query"
+	. "github.com/synnaxlabs/x/testutil"
 )
 
 var _ = Describe("Status", func() {
@@ -33,4 +40,34 @@ var _ = Describe("Status", func() {
 			Expect(v0.Status[any]{}.CustomTypeName()).To(Equal("Status"))
 		})
 	})
+})
+
+var _ = Describe("NormalizeKeys", func() {
+	It(
+		"Should lift a Status row stored under the pre-v0.54 key format",
+		func(ctx SpecContext) {
+			kvDB := memkv.New()
+			db := DeferClose(gorp.Wrap(kvDB, gorp.WithCodec(msgpack.Codec)))
+			e := v0.Status[any]{Key: "rack-1", Name: "Rack 1"}
+			legacy := gorptestutil.SetPreV54Row(
+				ctx,
+				kvDB,
+				"Status",
+				e.GorpKey(),
+				e,
+			)
+			table := MustOpen(
+				gorp.OpenTable(ctx, gorp.TableConfig[v0.Key, v0.Status[any]]{
+					DB:         db,
+					Migrations: []migrate.Migration{v0.NormalizeKeys},
+				}),
+			)
+			var res v0.Status[any]
+			Expect(table.NewRetrieve().
+				Where(gorp.MatchKeys[v0.Key, v0.Status[any]](e.GorpKey())).
+				Entry(&res).Exec(ctx, db)).To(Succeed())
+			Expect(res).To(Equal(e))
+			Expect(db.Get(ctx, legacy)).Error().To(MatchError(query.ErrNotFound))
+		},
+	)
 })
