+++ core/pkg/service/view/versions/v0/view_test.go
@@ -15,6 +15,13 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/synnaxlabs/synnax/pkg/service/ontology"
 	v0 "github.com/synnaxlabs/synnax/pkg/service/view/versions/v0"
+	"github.com/synnaxlabs/x/encoding/msgpack"
+	"github.com/synnaxlabs/x/gorp"
+	gorptestutil "github.com/synnaxlabs/x/gorp/testutil"
+	"github.com/synnaxlabs/x/kv/memkv"
+	"github.com/synnaxlabs/x/migrate"
+	"github.com/synnaxlabs/x/query"
+	. "github.com/synnaxlabs/x/testutil"
 )
 
 var _ = Describe("View", func() {
@@ -38,4 +45,32 @@ var _ = Describe("View", func() {
 			}))
 		})
 	})
+})
+
+var _ = Describe("NormalizeKeys", func() {
+	It(
+		"Should lift a View row stored under the pre-v0.54 key format",
+		func(ctx SpecContext) {
+			kvDB := memkv.New()
+			db := DeferClose(gorp.Wrap(kvDB, gorp.WithCodec(msgpack.Codec)))
+			e := v0.View{Key: uuid.New(), Name: "Recent", Type: "range"}
+			legacy := gorptestutil.SetPreV54Row(
+				ctx,
+				kvDB,
+				"View",
+				e.GorpKey(),
+				e,
+			)
+			table := MustOpen(gorp.OpenTable(ctx, gorp.TableConfig[v0.Key, v0.View]{
+				DB:         db,
+				Migrations: []migrate.Migration{v0.NormalizeKeys},
+			}))
+			var res v0.View
+			Expect(table.NewRetrieve().
+				Where(gorp.MatchKeys[v0.Key, v0.View](e.GorpKey())).
+				Entry(&res).Exec(ctx, db)).To(Succeed())
+			Expect(res).To(Equal(e))
+			Expect(db.Get(ctx, legacy)).Error().To(MatchError(query.ErrNotFound))
+		},
+	)
 })
